update kernel
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@65784 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/linux/universal/linux-4.4/net/batman-adv/bat_iv_ogm.c
@@ -450,7 +450,7 @@ static void batadv_iv_ogm_send_to_if(struct batadv_forw_packet *forw_packet,
 	struct batadv_priv *bat_priv = netdev_priv(hard_iface->soft_iface);
 	const char *fwd_str;
 	u8 packet_num;
-	s16 buff_pos;
+	int buff_pos;
 	struct batadv_ogm_packet *batadv_ogm_packet;
 	struct sk_buff *skb;
 	u8 *packet_pos;
